requirements.txt
tflite-runtime
opencv-python-headless
numpy
pillow
Flask
gunicorn